+++ crates/zng-ext-single-instance/src/lib.rs
@@ -190,9 +190,8 @@ event! {
 pub fn single_instance() {
     single_instance_named(
         std::env::current_exe()
+            .and_then(|p| p.canonicalize())
             .expect("current exe is required")
-            .canonicalize()
-            .expect("canonical current exe is required")
             .display()
             .to_txt(),
     )